modification edt
